Implement GTalk and Hangouts parsing.
GTalk and Hangouts have several awkward API differences, so it made more
sense to do two separate functions for them and concatenate them
together to get the final message list. I'm using the same content tag
for both so this should still ultimately treat them as the same source.

Note that for this to properly function the user needs to download all
their GTalk and Hangouts data in advance.

Author: Dylan Nugent